protect all table routes and add redis cache

+++ app/table/api.py
@@ -3,9 +3,10 @@ from typing import List
 from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi_cache.decorator import cache
 
+from app.auth.auth_backend import current_active_user
 from app.table.crud import get_all_tables, get_table, create_table, update_table, delete_table, get_tables
 from app.table.schema import TableResponse, TableCreate, TableUpdate
-from app.auth.database import get_async_session
+from app.auth.database import get_async_session, User
 
 router = APIRouter()
 
@@ -13,8 +14,11 @@ router = APIRouter()
 @router.post("/", response_model=TableResponse)
 async def create_table_endpoint(
         table: TableCreate,
-        db=Depends(get_async_session)
+        db=Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user.is_superuser:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     try:
         return await create_table(db, table)
     except Exception as e:
@@ -24,8 +28,11 @@ async def create_table_endpoint(
 @router.get("/all", response_model=List[TableResponse])
 @cache(expire=60)
 async def get_all_tables_endpoint(
-        db=Depends(get_async_session)
+        db=Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     try:
         return await get_all_tables(db)
     except Exception as e:
@@ -35,8 +42,11 @@ async def get_all_tables_endpoint(
 @router.get("/", response_model=List[TableResponse])
 @cache(expire=60)
 async def get_tables_endpoint(
-        db=Depends(get_async_session)
+        db=Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     try:
         return await get_tables(db)
     except Exception as e:
@@ -47,8 +57,11 @@ async def get_tables_endpoint(
 @cache(expire=60)
 async def get_table_endpoint(
         table_id: int,
-        db=Depends(get_async_session)
+        db=Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     try:
         return await get_table(db, table_id)
     except Exception as e:
@@ -59,8 +72,11 @@ async def get_table_endpoint(
 async def update_table_endpoint(
         table_id: int,
         table: TableUpdate,
-        db=Depends(get_async_session)
+        db=Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user.is_superuser:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     try:
         return await update_table(db, table_id, table)
     except Exception as e:
@@ -70,8 +86,11 @@ async def update_table_endpoint(
 @router.delete("/{table_id}")
 async def delete_table_endpoint(
         table_id: int,
-        db=Depends(get_async_session)
+        db=Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user.is_superuser:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     try:
         return await delete_table(db, table_id)
     except Exception as e:
